Initial commit of xorg-launch-helper.
There is not much left of uxlaunch OR user-session. After careful testing,
we can do almost all the bits that we need in systemd without any more
code than this.

The last remaining bit missing is Xorg startup synchronization. To start
a desktop session with X applications, we need to assure that Xorg is
ready before forking services that require an active $DISPLAY server.

This code comes all the way from uxlaunch and discards Xauth support, and
just includes the bits needed to start Xorg and wait for it to signal that
it is ready to accept and paint pixels for us. At that point, it exits
and thus tells systemd that the service is operational.

;###################################################################
;
;   THIS ENDS THE HISTORY FOR USER-SESSION.  :^)
;
;###################################################################