Attempt to fix shutdown issue causing 90sec timeout.
This adds a PIDFile for Xorg in the XDG_RUNTIME_DIR, in an attempt
to fix the shutdown timeout. The theory is that with this PIDFile,
system --user knows how to properly shut down Xorg, and won't
wait for 90 seconds to see it exit.

However, it doesn't work. I still see timeouts when shutting down
with this patch.