Convert to Type=notify. No hardcoded options.
This departs from the forking service, and rewrites xorg-launch-helper
to be a Notify type service to systemd. This allows us to monitor
and pass through TERM signals, as well as debug startup and shutdown
issues better.

Remove any harcoded options - any option needed to start Xorg now
needs to be passed to xorg-launch-helper as well. This can be done
through the unit file.